backend/requirements.txt
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==23.0.0
yfinance==0.2.40
pandas==2.2.0
numpy==1.26.3
requests==2.31.0

